pipeline_skeletton: move it to app and rename it as factory.py
